+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **_Breaking Change_**: remove `CBC.Add` and `UniStartBehaviour.Add`.
 - **_Breaking Change_**: `Entry.GetContext` will search parent first instead of the root object.
 - Feat: add interface `IReadonlyIoC`.
+- Feat: add `ComponentExtension.GetOrAddComponent`.
 - Fix: `AdvancedEvent` will call actions and once actions by the order they are added.
   - Previously all once actions will be called after repeated actions.
 - Fix: `AdvancedEvent` allow adding listeners or remove listeners during the invocation.